+++ src/main/java/umc/cockple/demo/domain/chat/AGENTS.md
@@ -13,7 +13,7 @@ Chat mixes REST queries with WebSocket transport, Redis-backed subscription/cach
 | WebSocket dispatch | `presentation/websocket/ChatWebSocketRequestDispatcher.java` | parses socket payloads and delegates authenticated requests |
 | WebSocket commands | `presentation/websocket/ChatWebSocketCommandHandler.java` | handles socket request types and publishes command events |
 | WebSocket responses | `presentation/websocket/WebSocketResponseSender.java` | serializes and sends socket responses |
-| WebSocket response assembly | `presentation/websocket/ChatWebSocketResponseAssembler.java` | shapes socket response payloads |
+| WebSocket response assembly | `converter/ChatWebSocketResponseAssembler.java` | shapes socket response payloads |
 | Realtime services | `service/websocket/` | subscription, room list cache, message fanout |
 | Read/query flows | `service/ChatQueryServiceImpl.java` | room lists, unread counts, history |
 | Events | `events/` | send/subscription events bridge transport and async handlers |
